Make it clearer how to free up the hardware serial for debugging.
On boards with only one hardware serial interface, some users connect the ESP8266 through the hardware serial because it provides the fastest connection to the ESP8266. However, by default this hardware serial port is used for logging. It can be freed up by defining a custom log facility (together with alternative definitions of LOGERROR, LOGDEBUG etc.).